refactor(datapack): URL host 정책 공유화
