+++ tools/datapack/validate-datapack.mjs
@@ -5,6 +5,7 @@ import { gunzipSync } from "node:zlib";
 import { DatabaseSync } from "node:sqlite";
 import { tmpdir } from "node:os";
 import path from "node:path";
+import { usesLocalPlaceholderHost } from "./production-url-policy.mjs";
 
 async function main() {
   const args = parseArgs(process.argv.slice(2));
@@ -812,102 +813,6 @@ function validateSourceInventory(sourceInventory, artifactKind, label) {
       }
     }
   }
-}
-
-function usesLocalPlaceholderHost(value) {
-  try {
-    let hostname = new URL(value).hostname.toLowerCase().replace(/\.$/, "");
-    if (hostname.startsWith("[") && hostname.endsWith("]")) {
-      hostname = hostname.slice(1, -1);
-    }
-    return isLocalPlaceholderHostname(hostname) || isNonPublicIpv4(hostname) || isNonPublicIpv6(hostname);
-  } catch {
-    return false;
-  }
-}
-
-function isLocalPlaceholderHostname(hostname) {
-  return hostname === "localhost" || hostname.endsWith(".localhost") || hostname.endsWith(".local");
-}
-
-function isNonPublicIpv4(hostname) {
-  const octets = parseIpv4Octets(hostname);
-  return octets !== null && isNonPublicIpv4Octets(octets);
-}
-
-function parseIpv4Octets(hostname) {
-  const parts = hostname.split(".");
-  if (parts.length !== 4) {
-    return null;
-  }
-  const octets = parts.map((part) => Number(part));
-  if (octets.some((octet) => !Number.isInteger(octet) || octet < 0 || octet > 255)) {
-    return null;
-  }
-  return octets;
-}
-
-function isNonPublicIpv4Octets(octets) {
-  const [first, second, third] = octets;
-  return (
-    first === 0 ||
-    first === 10 ||
-    (first === 100 && second >= 64 && second <= 127) ||
-    first === 127 ||
-    (first === 169 && second === 254) ||
-    (first === 172 && second >= 16 && second <= 31) ||
-    (first === 192 && second === 0 && (third === 0 || third === 2)) ||
-    (first === 192 && second === 88 && third === 99) ||
-    (first === 192 && second === 168) ||
-    (first === 198 && (second === 18 || second === 19)) ||
-    (first === 198 && second === 51 && third === 100) ||
-    (first === 203 && second === 0 && third === 113) ||
-    first >= 224
-  );
-}
-
-function isNonPublicIpv6(hostname) {
-  if (!hostname.includes(":")) {
-    return false;
-  }
-  const mappedIpv4 = parseIpv4MappedIpv6(hostname);
-  if (mappedIpv4 !== null) {
-    return isNonPublicIpv4Octets(mappedIpv4);
-  }
-  return (
-    hostname === "::" ||
-    hostname === "::1" ||
-    hostname.startsWith("fc") ||
-    hostname.startsWith("fd") ||
-    /^fe[89ab]/.test(hostname)
-  );
-}
-
-function parseIpv4MappedIpv6(hostname) {
-  if (!hostname.startsWith("::ffff:")) {
-    return null;
-  }
-  const suffix = hostname.slice("::ffff:".length);
-  if (suffix.includes(".")) {
-    return parseIpv4Octets(suffix);
-  }
-  const parts = suffix.split(":");
-  if (parts.length !== 2) {
-    return null;
-  }
-  const high = Number.parseInt(parts[0], 16);
-  const low = Number.parseInt(parts[1], 16);
-  if (
-    !Number.isInteger(high) ||
-    !Number.isInteger(low) ||
-    high < 0 ||
-    high > 0xffff ||
-    low < 0 ||
-    low > 0xffff
-  ) {
-    return null;
-  }
-  return [(high >> 8) & 0xff, high & 0xff, (low >> 8) & 0xff, low & 0xff];
 }
 
 function isAbsoluteHttpsWithHost(value) {
